Resume Placement type in Legend, Tooltip, and TransformControls and export top-level
