+++ packages/layerchart/src/lib/components/types.ts
@@ -1,0 +1,10 @@
+export type Placement =
+  | 'top-left'
+  | 'top'
+  | 'top-right'
+  | 'left'
+  | 'center'
+  | 'right'
+  | 'bottom-left'
+  | 'bottom'
+  | 'bottom-right';
